requirements.txt
# --- Core Functionality ---
python-dotenv
pypdf

# --- AI & Vector Store ---
sentence-transformers
httpx
# ChromaDB is an open-source embedding database, great for local development
chromadb
uuid

# --- FastAPI and Server ---
fastapi
uvicorn[standard]
pydantic

# --- LangChain for Utilities ONLY (used by ingestion_cli.py) ---
langchain-community
nltk

# --- Requests for ingestion_cli.py ---
requests